AT04Mflix/client/src/requirements.txt
apturl==0.5.2
asttokens==2.4.0
backcall==0.2.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
cliapp==1.20180812.1
click==8.0.3
cmdtest==0.32+git
colorama==0.4.4
comm==0.1.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
defer==1.0.6
distlib==0.3.7
distro==1.7.0
distro-info==1.1+ubuntu0.1
duplicity==0.8.21
exceptiongroup==1.1.3
executing==2.0.0
fasteners==0.14.1
filelock==3.13.0
future==0.18.2
grpcio==1.59.0
gyp==0.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
jeepney==0.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llfuse==1.3.8
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
monotonic==1.6
more-itertools==8.10.0
nest-asyncio==1.5.8
netifaces==0.11.0
numpy==1.26.1
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2023.10.24
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.12.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycups==2.0.1
Pygments==2.16.1
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.1.1
reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sklearn==0.0.post10
stack-data==0.6.3
systemd-python==234
tornado==6.3.3
traitlets==5.11.2
ttystatus==0.38
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.24.6
wadllib==1.3.6
wcwidth==0.2.8
xdg==5
xkit==0.0.0
zipp==1.0.0
